# internal/features/tasks/transport/http/patch_task.go
package tasks_transport_http

import (
	"net/http"

	"github.com/horizoonn/todoapp/internal/core/domain"
	core_logger "github.com/horizoonn/todoapp/internal/core/logger"
	core_http_request "github.com/horizoonn/todoapp/internal/core/transport/http/request"
	core_http_response "github.com/horizoonn/todoapp/internal/core/transport/http/response"
	core_http_types "github.com/horizoonn/todoapp/internal/core/transport/http/types"
)

type PatchTaskRequest struct {
	Title       core_http_types.Nullable[string] `json:"title"         swaggertype:"string" example:"Walk the dog"`
	Description core_http_types.Nullable[string] `json:"description"   swaggertype:"string" example:"null"`
	Completed   core_http_types.Nullable[bool]   `json:"completed"     swaggertype:"boolean"`
}

type PatchTaskResponse TaskDTOResponse

// PatchTask     godoc
// @Summary      Обновить задачу
// @Description  Обновляет информацию об уже существующей в системе задаче
// @Description  ### Логика обновления полей (Three-state logic):
// @Description  1. **Поле не передано**: `description` игнорируется, значение в БД не меняется
// @Description  2. **Явно передано значение**: `"description": "Утром в 06:30 выйти на прогулку с Бобиком"` — устанавливает новое описание для задачи
// @Description  3. **Явно передан null**: `"description": null`— очищает поле в БД (set to NULL)
// @Description  Ограничения: `title` и `completed` не могут быть выставлены как null
// @Tags         tasks
// @Accept       json
// @Produce      json
// @Param        id       path      string               true  "ID изменяемой задачи" Format(uuid)
// @Param        request  body      PatchTaskRequest  true  "PatchTask тело запроса"
// @Success      200      {object}  PatchTaskResponse                "Успешно изменённая задача"
// @Failure      400      {object}  core_http_response.ErrorResponse "Bad request"
// @Failure      404      {object}  core_http_response.ErrorResponse "Task not found"
// @Failure      409      {object}  core_http_response.ErrorResponse "Conflict"
// @Failure      500      {object}  core_http_response.ErrorResponse "Internal server error"
// @Router       /tasks/{id} [patch]
func (h *TasksHTTPHandler) PatchTask(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	taskID, err := core_http_request.GetUUIDPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to get taskID path value")

		return
	}

	var request PatchTaskRequest

	if err := core_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(err, "failed to decode and validate HTTP request")

		return
	}

	taskPatch := taskPatchFromRequest(request)

	taskDomain, err := h.tasksService.PatchTask(ctx, taskID, taskPatch)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to patch task")

		return
	}

	response := PatchTaskResponse(taskDTOFromDomain(taskDomain))

	responseHandler.JSONResponse(response, http.StatusOK)
}

func taskPatchFromRequest(request PatchTaskRequest) domain.TaskPatch {
	return domain.NewTaskPatch(
		request.Title.ToDomain(),
		request.Description.ToDomain(),
		request.Completed.ToDomain(),
	)
}